pyproject.toml
[tool.poetry]
name = "file_storage"
version = "0.1.0"
description = ""
authors = ["Nikita Kolybelkin <kennedywhokilled@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.7.3"
yoyo-migrations = "^7.2.1"
asyncpg = "^0.21.0"
pydantic = "^1.7.3"
psycopg2-binary = "^2.8.6"
pytest-aiohttp = "^0.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "^20.8b1"
isort = "^5.6.4"

[tool.isort]
line_length = "79"
multi_line_output = "3"
include_trailing_comma = "true"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "FIRSTPARTY"
known_first_party = "file_storage"

[tool.black]
line-length = 79
target_version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
    | \.venv
  )/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
